requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-Limiter==4.0.0
Flask-Talisman==1.1.0
Flask-WTF==1.2.1
Flask-Caching==2.3.1
Werkzeug==3.1.3
Jinja2==3.1.5
WTForms==3.2.1
email-validator==2.1.0
python-dotenv==1.0.1
bleach==6.2.0
tinycss2==1.3.0
markdown==3.7
gunicorn==23.0.0
google-generativeai==0.7.2
Flask-Mail==0.9.1
pyotp==2.9.0
qrcode[pil]==7.4.2
cryptography==41.0.7
redis==5.2.0
sentry-sdk[flask]==2.18.0
psycopg2-binary==2.9.10
pyodbc==5.1.1
